apps/api-py/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
supabase==2.6.0
python-dotenv==1.0.1
PyJWT==2.10.1
